pyproject.toml
[project]
name = "memex"
version = "0.2.1"
description = "Kiki's personal blog automation suite"
authors = [{name = "Kiki", email = "kiki@example.com"}]
requires-python = ">=3.9"
dependencies = ["click", "PyYAML", "rich"]

[project.scripts]
memex = "memex.cli:cli"

[build-system]
requires      = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where    = [""]       
include  = ["memex*"] 
exclude  = ["tests*", "docs"]